feat: integrate the data access from oauth to sign up process page
